Change to version dependence
the new release of qiime2.2019-10 depends on the package SEPP.  That package package has a dependency for `hmmer==3.1b2`.  itsxpress has a dependency for `hmmer >=3.1` but the way conda parses versions it thinks 3.1b2 is less than 3.1.  See the issue here: 
 https://github.com/USDA-ARS-GBRU/itsxpress/issues/16

Changing the dependency in my build to `hmmer>=3.1b2` should solve this.